use `addSegment` instead of `addPath` for http4s
When building a http4s request, currently `addPath` is used to append segments to the path. Which results in the string being split by `/` and added as multiple segments (URL-encoding them).

So right now it's impossible to have a path segment that contains a `/` symbol (an edge case, but still):
* if you pre-url-encode it – http4s will encode it again, and you get the double encoded `/`: `/` -> `%2f` -> `%252f`
* if you don't url-encode it – it ends up "unencoded" in the resulting string, causing a segment split

In either case, you can't get the desired `%2f` encoding.