Codegen: generate servers object (#4492)
Generating code from the 'servers' section of the yaml.

Also adds 'name' field to endpoint definitions.

I may have missed something that already exists in the tapir model that
I should've been using here, although I didn't see it in my (admittedly
not very exhaustive) look. I had a go at producing something that could
actually be useful --- the idea is essentially that if the openapi
schema defined something for the servers section, a client has a decent
chance of getting a valid URI by just digging into the configuration and
using the default values.

I'm not certain if this is actually a good idea or not, but it seems
useful...

If the generated server defns here feel a bit wrong, but we still want
the 'name' parameter defined, I'm happy to reduce the pr to just include
the latter, although ideally I do kinda want those default server URLs
defined somewhere at least.

Even if this is deemed a reasonable approach, it's also an open question
in my mind whether we really want those uris to be strings... Possibly
would be better to construct an sttp.model.Uri instead.