codegen: Support security path declarations (#4516)
Adds a new custom extension 'x-tapir-codegen-security-path-prefixes',
for when your auth logic wants to know some prefix of the path (e.g. for
validating that a jwt contains perms for the segment)

The path security input will be prepended after any other security
inputs are declared and mapped, so that the security path will appear
before other security inputs in the signature. This matches the
non-security inputs, where path is also first.

There may be a neater way to do this that I've missed, but my
understanding is that security inputs are _always_ parsed before
non-security input, so if you want to know about the path in the
security logic this is the only approach that works...