codegen(sbt): handle cases where default openapi file does not exist (#4523)
Addresses https://github.com/softwaremill/tapir/issues/4388

In order to achieve 'least surprise', this handles two cases more
gracefully -- the case where the 'default' openapi file is also
referenced in the 'openapiAdditionalPackages' setting (in which case the
packaged declared for the latter will win, and it will not be
duplicated), and the case where 'openapiAdditionalPackages' is non-empty
and the default file is entirely missing, in which case we will assume
the default file is undesired and not error.

In both cases a warning will be emitted to stdErr

It will still be an error if the default openapi file is missing _and_
the 'openapiAdditionalPackages' setting is empty. In which case no valid
files have been found and erroring is appropriate.