Document semi-automatic schema derivation limitation with generic classes
When doing semi-automatic schema derivation on generic classes, the name
of the generated schema does not use the specific type parameter. This
leads to producing inconsistent OpenAPI specification in some cases.

This issue appears to be quite hard to fix, due to the way Magnolia
derivation is implemented (cf.
https://github.com/softwaremill/tapir/issues/3922#issuecomment-2841929069).
Instead, document this issue as a limitation, and document a possible
work-around using `inline given`.

Fixes: https://github.com/softwaremill/tapir/issues/3922
Fixes: https://github.com/softwaremill/tapir/issues/4549