Codegen: support validation (#4564)
The goal here is to support validation at a fairly comprehensive level,
by constructing recursive validators (where applicable), allowing us to
apply logic even onto leaf fields in nested trees (at least for
json/xml/formData).

Since validation may not be desirable, and since the validator
construction could break a spec that would otherwise generate correct
code, I've added a new flag to permit disabling validation, but it's on
by default because that seems a preferable default (and maintaining
_strict_ backwards compatibility on a plugin doesn't seem very pressing
to me...). I have tried to ensure the test cases are fairly exhaustive,
so as to minimise the chance that this will actually break anything in a
real spec.

This pr doesn't quite yet support all the validation we might want.
Exclusions are:
- uniqueItems (this would be 'easy' to support as an assertion, but I'm
not yet sure that's what we really want -- perhaps we want to generate
'Set' rather than 'Seq' schemas for that?)
- readOnly/writeOnly
- minProperties/maxProperties on 'object' schemas (although this _is_
supported on 'map' schemas, which I should imagine is the normal
usecase)