codegen: support singleton allOf for default on refs (#4604)
A use-case for allOf that I'd missed is supporting default values on
fields whose type are declared by 'ref's. In fact, there are really the
only two ways to do that that works with openapi 3.0:
```yaml
allOf:
 - $ref: ...someRef
 - default: defaultValue
```
and

```yaml
allOf:
 - $ref: ...someRef
default: defaultValue
```
since sibling fields to $ref are ignored.

I'm only supporting the latter form here, and _not_ attempting to
support other usages of allOf outside of top-level usage in
components/schemas definitions (which are already largely handled). This
is really a very specific case, and more general handling would be _far_
more involved

(github's default diff on this looks really weird, it makes more sense
with 'hide whitespace' enabled)