codegen: make generated class field & endpoint decl orders consistent with api declarations (#4605)
The non-deterministic ordering of the field declarations in object
schemas is annoying, and can randomly break code that depends on it
being consistent (e.g. case-class pattern matching, or positional case
class construction). The sanest consistent field ordering is the one
declared in the openapi yaml. This pr replaces usage of Map with
LinkedHashMap, which retains the initial ordering of fields as declared
in the schema.

Where 'allOf' is involved, fields will be declared in the order in which
they first appear in the flattened property list of all component
schemas.

Endpoint ordering has also been non-deterministic, which can also be
problematic (although it's probably less of an issue for most). That has
a similar resolution - to read the paths into a LinkedHashMap rather
than a Map, so as to preserve declaration ordering.

Honestly, this should've been done ages ago, since 'fixing' the ordering
does have the potential to break code on plugin bump -- although
changing the type of a field (or whether it has a default) can already
randomly break stuff, the plugin version can be updated entirely
independently of other updates, and better to do this late than never I
guess...