codegen: handle recursive schema defns in scala 3 (#4671)
Supports deriving schemas for recursive schema definitions in scala 3.

Without this change, the code will hang eternally at runtime. This
appears to be because, when declared as a `lazy val`, the tapir schema
for a recursive struct enters some infinite loop. `def` declarations
have no such issues; however `lazy val` is, probably, preferable for
most cases becase the schema only needs to be evaluated once, even if
used in multiple locations - so there's a little extra complexity in the
generation now, so that we only declare the schemas as `def`s when
they're recursive (or part of a mutually-recursive set of defns, which
had the same issue).

I guess it'd be nicer to fix the 'root issue' with lazy val schema
declarations, but that's probably some magnolia issue that I have not
the wit to fix, and this approach does well enough.