codegen: Add directives to force body serialisation to eager or streaming (#4686)
The default behaviours for endpoints are not always desirable --
sometimes the complexity of streaming for a text/csv response body are
overkill, and sometimes you really want to stream json in. This pr adds
`force-eager` and `force-streaming` options to the per-endpoint custom
directives, so that default behaviour can be overridden.

There are other variants for specifying behaviour of just req or just
resp bodies, for a total of 6 new directives.

Errors are still always eager.