Remove reliance of tests on Cats Effect, to enable Scala Native release for sttp client module (#4617)
Most tests operate within or via Future anyways, so we remove the level
of indirection and express as much as possible in terms of Future,
removing the need for CE dependency in a lot of places, but especially
on Native. I got `sttpClient4Native4/test` to pass locally, which is all
I care about really.

Note that CE release on Native 0.5 is imminent, but if it's delayed,
this might be _an_ option to get native builds of sttp modules out (and
may be others?).

Getting a native tapir-sttp-client module is my only goal here, so I
don't care if it's achieved by any other means – you can consider this
PR if you think it has merit on its own.

not a Draft as I want to get a CI run out of it.

---------

Co-authored-by: Adam Warski <adam@warski.org>