Implement multipart support for Netty Sync server (#4849)
This PR is based on #3933 and implements support for multipart
requests/responses with Netty. Currently only for Netty Sync, I will try
to implement for Cats, ZIO, etc. in another PR.

Adds a dependency on the `httpmime` library for using the
`MultipartEntityBuilder` (as used in JDK HTTP server), since Netty
unfortunately has no built-in way to make multipart response bodies.

(I have not included the original commits from #3933, but if that's
desired I can do some rebasing to include them!)