openapi: Fix mutual-recursion bug in scala 2.13 (issue 4911) (#4916)
Contains the fix for a few disparate things:

- fix runtime stackoverflow for mutually-recursive scheme definitions in
scala 2.13 ([issue](https://github.com/softwaremill/tapir/issues/4911))
- permit defining schema aliases for simple types (i.e
   ```yaml
     ClientIdParameterStr:
      type: string
   ```
    becomes `type ClientIdParameterStr = String`)
      
- handle missing descriptions on responses (these are technically
required by the spec, but often absent IRL)
- handle non-200 streaming responses with descriptions 
- handle optional streaming responses with empty alternatives defined
for other status codes
- produce io.circe.JsonObject for [free-form
object](https://swagger.io/docs/specification/v3_0/data-models/data-types/#free-form-object)
types, rather than empty case classes (note that old behaviour can be
achieved by adding an empty `properties` map to the definition; but that
an empty case class was never an appropriate model to generate for a
free form object). This is not always going to be the correct model --
arguably, for any non-json endpoint, it's probably not ideal -- but it's
the most generic model that's known to be available, and which can be
generally mapped to object schema definitions.
- handle inline enum definitions in paths
- default `*/*` to strict for convenience (this behaviour can be
overridden by the appropriate directives, but seems like it's _probably_
a saner default)

Includes the swagger from
https://nifi.apache.org/nifi-docs/swagger.yaml, which prompted these
fixes - since it's an OSS apache project, that seems legit - along with
a few modifications to it (mostly commenting out some security
declarations, although also removes some 'object' wrappers around what
would otherwise be primitive schema declarations, fudges the wrapper
around `BulletinBoardPatternParameter` a bit since doing otherwise would
require supporting objects in query params, and adds a `flows` field to
the `HTTPBearerJWT` decl. Doesn't address the fact that some query
params look like they should probably be arrays judging by the
description, nor does it try to make sense of the security declarations
since they don't match up with any openapi conventions I'm aware of).

This is a pretty chunky spec and adds a couple of mins to the openapi
test duration, but I think it's worth it to get the extra coverage...

Given the nature of the bug, we not only run the check to validate
expected scala output and that the code compiles, but also explicitly
check that the `controllerServiceReferencingComponentDTOTapirSchema`
schema doesn't throw a `StackOverflow` at runtime.