More precise metrics & logging (#4943)
In some cases, responses are generated by interceptors: 404 rejections
using `RejectInterceptor`, CORS preflight requests, not-acceptable
interceptor, exception interceptor. In most of these cases, there are no
`Endpoint`s that are associated with the requests.

Whatever the source of the response (endpoint invocation, decode
failure, 404 rejection), such responses should be handled by
interceptors such as logging & metrics.

To properly log such interceptor-generated (or more precisely:
`RequestHandler`-generated, without invoking an `EndpointHandler`)
responses, the log interceptor must know if the response was generated
by an endpoint, or request handler. This is achieved by adding an
additional field to `RequestResult.Response`.

The logging interceptor is enhanced by calling a new
`ServerLog.requestHandledByInterceptor` method, which should log the
interceptor-generated response. Moreover, the exception & logging
interceptors are combined, so that upon an exception, first an exception
is logged, then the exception is turned into a 500 response, and then
that response is logged as well.

The metrics interceptor is enhanced so that when a request is received,
there's always some callback on the `EndpointMetric` (returned by the
`Metric`) that is called - including interceptor-generated responses.

Finally, active-requests gauge/up-down counter is changed so that it
counts all requests. This also means that there's no endpoint-template
(path) label associated with the metric.

The ordering of interceptors is changed. Now, the metrics &
logging+exception interceptors come first, followed by CORS, reject &
not-acceptable.